Provides banking security Through advanced neural networks
By introducing a novel security paradigm that gets around the drawbacks of biometrics and video monitoring, this research solves shortcomings in conventional banking access control. To improve intrusion detection and tighten access control, the system combines neural net, google cloud, and network-messaging applications. The core is a multi-layered security system, where the first layer starts with picture capture. This layer uses transfer learning to validate people, allowing access to only verified humans. Convolutional Neural Networks (CNNs) are utilised in the second layer to process photos and verify the individual's identity as the locker owner. The image is instantly sent via cloud and network applications to the owner's device upon access denial. Users need to input a password at the third layer, and if they enter it correctly, an OTP is generated. After three failed tries at entering the password, the system alerts the owner with an OTP that contains the image that was collected. Opening the locker is part of the fourth layer, which is constantly watched. Strong security mechanisms are used by each layer, such as CNNs, transfer learning, and smooth network and cloud integration. This all-encompassing security solution, which is implemented using a web framework built on Python, is successful in strengthening access control at every level.